testing/web-platform/tests/dom/nodes/Document-createElementNS.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Document.createElementNS</title>
<link rel=help href="http://dom.spec.whatwg.org/#dom-document-createelementns">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  var invalidNames = [
    "",
    "1foo",
    "\u0300foo",
    "}foo",
    "f}oo",
    "foo}",
    "\ufffffoo",
    "f\uffffoo",
    "foo\uffff",
    "<foo",
    "foo>",
    "<foo>",
    "f<oo"
  ],
  invalidNSQNameCombinations = [
    ["", ":foo"],
    ["", "foo:"],
    ["", "foo:foo"],
    ["http://oops/", "xml:foo"],
    ["http://oops/", "xmlns"],
    ["http://oops/", "xmlns:foo"],
    ["http://www.w3.org/2000/xmlns/", "xml:foo"],
    ["http://www.w3.org/2000/xmlns/", "foo:xmlns"]
  ];

  invalidNames.forEach(function(name) {
    test(function() {
      assert_throws("INVALID_CHARACTER_ERR", function() { document.createElementNS("", name) });
    }, "Invalid Name: " + name)
  })

  invalidNSQNameCombinations.forEach(function(a) {
    test(function() {
      assert_throws("NAMESPACE_ERR", function() { document.createElementNS(a[0], a[1]) });
    }, "Invalid namespace " + a[0] + " and QName " + a[1])
  })
})

test(function() {
  var HTMLNS = "http://www.w3.org/1999/xhtml";
  var element = document.createElementNS(HTMLNS, "span");
  assert_equals(element.namespaceURI, HTMLNS);
  assert_equals(element.prefix, null);
  assert_equals(element.localName, "span");
  assert_equals(element.tagName, "SPAN");
  assert_true(element instanceof Node, "Should be a Node");
  assert_true(element instanceof Element, "Should be a Element");
  assert_true(element instanceof HTMLElement, "Should be a HTMLElement");
  assert_true(element instanceof HTMLSpanElement, "Should be a HTMLSpanElement");
}, "Lower-case HTML element without a prefix");

test(function() {
  var HTMLNS = "http://www.w3.org/1999/xhtml";
  var element = document.createElementNS(HTMLNS, "html:span");
  assert_equals(element.namespaceURI, HTMLNS);
  assert_equals(element.prefix, "html");
  assert_equals(element.localName, "span");
  assert_equals(element.tagName, "HTML:SPAN");
  assert_true(element instanceof Node, "Should be a Node");
  assert_true(element instanceof Element, "Should be a Element");
  assert_true(element instanceof HTMLElement, "Should be a HTMLElement");
  assert_true(element instanceof HTMLSpanElement, "Should be a HTMLSpanElement");
}, "Lower-case HTML element with a prefix");

test(function() {
  var element = document.createElementNS("test", "span");
  assert_equals(element.namespaceURI, "test");
  assert_equals(element.prefix, null);
  assert_equals(element.localName, "span");
  assert_equals(element.tagName, "span");
  assert_true(element instanceof Node, "Should be a Node");
  assert_true(element instanceof Element, "Should be a Element");
  assert_false(element instanceof HTMLElement, "Should not be a HTMLElement");
  assert_false(element instanceof HTMLSpanElement, "Should not be a HTMLSpanElement");
}, "Lower-case non-HTML element without a prefix");

test(function() {
  var element = document.createElementNS("test", "html:span");
  assert_equals(element.namespaceURI, "test");
  assert_equals(element.prefix, "html");
  assert_equals(element.localName, "span");
  assert_equals(element.tagName, "html:span");
  assert_true(element instanceof Node, "Should be a Node");
  assert_true(element instanceof Element, "Should be a Element");
  assert_false(element instanceof HTMLElement, "Should not be a HTMLElement");
  assert_false(element instanceof HTMLSpanElement, "Should not be a HTMLSpanElement");
}, "Lower-case non-HTML element with a prefix");

test(function() {
  var HTMLNS = "http://www.w3.org/1999/xhtml";
  var element = document.createElementNS(HTMLNS, "SPAN");
  assert_equals(element.namespaceURI, HTMLNS);
  assert_equals(element.prefix, null);
  assert_equals(element.localName, "SPAN");
  assert_equals(element.tagName, "SPAN");
  assert_true(element instanceof Node, "Should be a Node");
  assert_true(element instanceof Element, "Should be a Element");
  assert_true(element instanceof HTMLElement, "Should be a HTMLElement");
  assert_true(element instanceof HTMLUnknownElement, "Should be a HTMLUnknownElement");
  assert_false(element instanceof HTMLSpanElement, "Should not be a HTMLSpanElement");
}, "Upper-case HTML element without a prefix");

test(function() {
  var HTMLNS = "http://www.w3.org/1999/xhtml";
  var element = document.createElementNS(HTMLNS, "html:SPAN");
  assert_equals(element.namespaceURI, HTMLNS);
  assert_equals(element.prefix, "html");
  assert_equals(element.localName, "SPAN");
  assert_equals(element.tagName, "HTML:SPAN");
  assert_true(element instanceof Node, "Should be a Node");
  assert_true(element instanceof Element, "Should be a Element");
  assert_true(element instanceof HTMLElement, "Should be a HTMLElement");
  assert_false(element instanceof HTMLSpanElement, "Should not be a HTMLSpanElement");
}, "Upper-case HTML element with a prefix");

test(function() {
  var element = document.createElementNS("test", "SPAN");
  assert_equals(element.namespaceURI, "test");
  assert_equals(element.prefix, null);
  assert_equals(element.localName, "SPAN");
  assert_equals(element.tagName, "SPAN");
  assert_true(element instanceof Node, "Should be a Node");
  assert_true(element instanceof Element, "Should be a Element");
  assert_false(element instanceof HTMLElement, "Should not be a HTMLElement");
  assert_false(element instanceof HTMLSpanElement, "Should not be a HTMLSpanElement");
}, "Upper-case non-HTML element without a prefix");

test(function() {
  var element = document.createElementNS("test", "html:SPAN");
  assert_equals(element.namespaceURI, "test");
  assert_equals(element.prefix, "html");
  assert_equals(element.localName, "SPAN");
  assert_equals(element.tagName, "html:SPAN");
  assert_true(element instanceof Node, "Should be a Node");
  assert_true(element instanceof Element, "Should be a Element");
  assert_false(element instanceof HTMLElement, "Should not be a HTMLElement");
  assert_false(element instanceof HTMLSpanElement, "Should not be a HTMLSpanElement");
}, "Upper-case non-HTML element with a prefix");
</script>
